% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_kmer_scores}
\alias{calculate_kmer_scores}
\title{\emph{k}-mer Score Calculation}
\usage{
calculate_kmer_scores(kmers, pwm)
}
\arguments{
\item{kmers}{list of \emph{k}-mers}

\item{pwm}{position weight matrix}
}
\value{
list of PWM scores for the specified \emph{k}-mers
}
\description{
C++ implementation of \emph{k}-mer score calculation
}
\examples{
motif <- get_motif_by_id("M178_0.6")[[1]]
kmers <- c("AAAAAA", "CAAAAA", "GAAAAA")
calculate_kmer_scores(kmers, as.matrix(motifMatrix(motif)))

}
