% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{calculateKmerEnrichment}
\alias{calculateKmerEnrichment}
\title{\emph{k}-mer Enrichment between Foreground and Background Sets}
\usage{
calculateKmerEnrichment(
  foreground_sets,
  background_set,
  k,
  permutation = FALSE,
  chisq_p_value_threshold = 0.05,
  p_adjust_method = "BH",
  n_cores = 4
)
}
\arguments{
\item{foreground_sets}{list of foreground sets; a foreground set is a
character vector of
DNA or RNA sequences (not both) and a strict subset of the
\code{background_set}}

\item{background_set}{character vector of DNA or RNA sequences that
constitute the
background set}

\item{k}{length of \emph{k}-mer, either \code{6} for hexamers or
\code{7} for heptamers}

\item{permutation}{if \code{TRUE}, only the enrichment value is returned
(efficiency mode
used for permutation testing)}

\item{chisq_p_value_threshold}{threshold below which Fisher's exact test
is used instead of Pearson's chi-squared test}

\item{p_adjust_method}{see \code{\link[stats]{p.adjust}}}

\item{n_cores}{number of computing cores to use}
}
\value{
A list with two entries:

(1) dfs: a list of data frames with results from
\code{\link{computeKmerEnrichment}} for each of the foreground sets
(2) kmers: a character vector of all k-mers
}
\description{
Calls \code{\link{computeKmerEnrichment}} to compute \emph{k}-mer
enrichment values
for multiple foregrounds. Calculates enrichment for foreground sets in
parallel.
}
\examples{
# define simple sequence sets for foreground and background
foreground_set1 <- c(
  "CAACAGCCUUAAUU", "CAGUCAAGACUCC", "CUUUGGGGAAU",
  "UCAUUUUAUUAAA", "AAUUGGUGUCUGGAUACUUCCCUGUACAU",
  "AUCAAAUUA", "AGAU", "GACACUUAAAGAUCCU",
  "UAGCAUUAACUUAAUG", "AUGGA", "GAAGAGUGCUCA",
  "AUAGAC", "AGUUC", "CCAGUAA"
)
foreground_set2 <- c("UUAUUUA", "AUCCUUUACA", "UUUUUUU", "UUUCAUCAUU")
foreground_sets <- list(foreground_set1, foreground_set2)
background_set <- c(foreground_set1, foreground_set2,
                    "CCACACAC", "CUCAUUGGAG", "ACUUUGGGACA", "CAGGUCAGCA")

# single-threaded
kmer_enrichment_values_st <- calculateKmerEnrichment(foreground_sets,
  background_set, 6, n_cores = 1)
\dontrun{
# multi-threaded
kmer_enrichment_values_mt <- calculateKmerEnrichment(foreground_sets,
  background_set, 6)}
}
\seealso{
Other \emph{k}-mer functions: 
\code{\link{checkKmers}()},
\code{\link{computeKmerEnrichment}()},
\code{\link{drawVolcanoPlot}()},
\code{\link{empiricalEnrichmentMeanCDF}()},
\code{\link{generateKmers}()},
\code{\link{generatePermutedEnrichments}()},
\code{\link{homopolymerCorrection}()},
\code{\link{permTestGeometricMean}()},
\code{\link{runKmerSPMA}()},
\code{\link{runKmerTSMA}()}
}
\concept{\emph{k}-mer functions}
