% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateLocalConsistency}
\alias{calculateLocalConsistency}
\title{Local Consistency Score}
\usage{
calculateLocalConsistency(x, numPermutations, minPermutations, e)
}
\arguments{
\item{x}{numeric vector that contains values for shuffling}

\item{numPermutations}{maximum number of permutations performed in Monte Carlo test
for consistency score}

\item{minPermutations}{minimum number of permutations performed in Monte Carlo test
for consistency score}

\item{e}{stop criterion for consistency score Monte Carlo test: aborting permutation
process after observing \code{e} random consistency values with more extreme values
than the actual consistency value}
}
\description{
C++ implementation of Local Consistency Score algorithm.
}
