% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateTranscriptMC}
\alias{calculateTranscriptMC}
\title{Motif Enrichment calculation}
\usage{
calculateTranscriptMC(absoluteHits, totalSites, relHitsForeground, n,
  maxPermutations, minPermutations, e)
}
\arguments{
\item{absoluteHits}{number of putative binding sites per sequence
(returned by \code{\link{scoreTranscripts}})}

\item{totalSites}{number of potential binding sites per sequence
(returned by \code{\link{scoreTranscripts}})}

\item{relHitsForeground}{relative number of hits in foreground set}

\item{n}{number of sequences in the foreground set}

\item{maxPermutations}{maximum number of foreground permutations
performed in
Monte Carlo test for enrichment score}

\item{minPermutations}{minimum number of foreground permutations
performed in
Monte Carlo test for enrichment score}

\item{e}{stop criterion for enrichment score Monte Carlo test:
aborting permutation process
after observing \code{e} random enrichment values with more extreme
values than the actual
enrichment value}
}
\value{
list with p-value and number of iterations of Monte Carlo sampling
for foreground enrichment
}
\description{
C++ implementation of Motif Enrichment calculation
}
\examples{
foreground.seqs <- c("CAGUCAAGACUCC", "AAUUGGUUGUGGGGCUUCCCUGUACAU",
                     "AGAU", "CCAGUAA", "UGUGGGG")
background.seqs <- c(foreground.seqs, "CAACAGCCUUAAUU", "CUUUGGGGAAU",
                     "UCAUUUUAUUAAA", "AUCAAAUUA", "GACACUUAAAGAUCCU",
                     "UAGCAUUAACUUAAUG", "AUGGA", "GAAGAGUGCUCA",
                     "AUAGAC", "AGUUC")
motif.db <- getMotifById("M178_0.6")
fg <- scoreTranscripts(foreground.seqs, cache = FALSE,
  motifs = motif.db)
bg <- scoreTranscripts(background.seqs, cache = FALSE,
  motifs = motif.db)

mc.result <- calculateTranscriptMC(unlist(bg$absolute.hits),
 unlist(bg$total.sites),
 fg$df$absolute.hits / fg$df$total.sites,
 length(foreground.seqs), 1000, 500, 5)
}
