% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{checkKmers}
\alias{checkKmers}
\title{Check Validity of Set of \emph{k}-mers}
\usage{
checkKmers(kmers)
}
\arguments{
\item{kmers}{set of \emph{k}-mers}
}
\value{
\code{TRUE} if set of \emph{k}-mers is valid
}
\description{
Checks if the provided set of \emph{k}-mers is valid. A valid set
of \emph{k}-mers is
(1) non-empty, (2) contains either only hexamers or only heptamers,
 and (3) contains
only characters from the RNA alphabet (A, C, G, U)
}
\examples{
# valid set
checkKmers(c("ACGCUC", "AAACCC", "UUUACA"))

# invalid set (contains hexamers and heptamers)
checkKmers(c("ACGCUC", "AAACCC", "UUUACAA"))
}
\seealso{
Other \emph{k}-mer functions: \code{\link{calculateKmerEnrichment}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{drawVolcanoPlot}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{\emph{k}-mer functions}
