% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ge}
\alias{ge}
\title{Toy Gene Expression Data Set}
\format{
A list with the following components:
\tabular{rl}{
  \code{foreground1_df} \tab data frame that contains down-regulated
  genes after treatment\cr
  \code{foreground2_df} \tab data frame that contains up-regulated
  genes after treatment\cr
  \code{background_df} \tab data frame that contains all genes measured
}
}
\usage{
ge
}
\description{
This object contains a toy data set based on gene expression measurements
and 3'-UTR sequences of 1000 genes. It comprises three data frames with
RefSeq identifiers, log fold change values, and 3'-UTR sequences of genes,
which are either upregulated or downregulated after some hypothetical
treatment, as well as all measured genes. The actual values are not
important. This data set merely serves as an example input for various
functions.
}
\keyword{datasets}
