% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{generateIUPACByKmers}
\alias{generateIUPACByKmers}
\title{Generates IUPAC code for a character vector of \emph{k}-mers}
\usage{
generateIUPACByKmers(kmers, code = NULL)
}
\arguments{
\item{kmers}{character vector of \emph{k}-mers}

\item{code}{if IUPAC code table has already been initialized by
\code{\link{initIUPAClookupTable}}, it can be
specified here}
}
\value{
the IUPAC string of the binding site
}
\description{
Generates a compact logo of a motif based on IUPAC codes given by a character
vector of \emph{k}-mers
}
\details{
IUPAC RNA nucleotide code:
\tabular{rl}{
  \code{A} \tab Adenine\cr
  \code{C} \tab Cytosine\cr
  \code{G} \tab Guanine\cr
  \code{U} \tab Uracil\cr
  \code{R} \tab A or G\cr
  \code{Y} \tab C or U\cr
  \code{S} \tab G or C\cr
  \code{W} \tab A or U\cr
  \code{K} \tab G or U\cr
  \code{M} \tab A or C\cr
  \code{B} \tab C or G or U\cr
  \code{D} \tab A or G or U\cr
  \code{H} \tab A or C or U\cr
  \code{V} \tab A or C or G\cr
  \code{N} \tab any base
}
}
\examples{
generateIUPACByKmers(c('AACCAA', 'AACCGG', 'CACCGA'))
}
\references{
\url{http://www.chem.qmul.ac.uk/iubmb/misc/naseq.html}
}
\seealso{
Other motif functions: \code{\link{generateIUPACByMatrix}},
  \code{\link{generateKmersFromIUPAC}},
  \code{\link{getMotifById}}, \code{\link{getMotifByRBP}},
  \code{\link{getMotifs}}, \code{\link{getPPM}},
  \code{\link{initIUPAClookupTable}},
  \code{\link{motifsMetaInfo}}, \code{\link{setMotifs}}
}
\concept{motif functions}
