% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{geometricMean}
\alias{geometricMean}
\title{Geometric Mean}
\usage{
geometricMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector of values for which the geometric mean will be computed}

\item{na.rm}{logical.  Should missing values (including \code{NaN}) be
    removed?}
}
\value{
Geometric mean of \code{x} or \code{1} if length of \code{x} is 0
}
\description{
Calculates the geometric mean of the specified values.
}
\examples{
geometricMean(c(0.123, 0.441, 0.83))

}
