% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_motifs}
\alias{get_motifs}
\title{Retrieve list of all motifs}
\usage{
get_motifs()
}
\value{
A list of objects of class Motif
}
\description{
Retrieves all Transite motifs
}
\examples{
transite_motifs <- get_motifs()
}
\seealso{
Other motif functions: 
\code{\link{generate_iupac_by_kmers}()},
\code{\link{generate_iupac_by_matrix}()},
\code{\link{generate_kmers_from_iupac}()},
\code{\link{get_motif_by_id}()},
\code{\link{get_motif_by_rbp}()},
\code{\link{get_motifs_meta_info}()},
\code{\link{get_ppm}()},
\code{\link{init_iupac_lookup_table}()},
\code{\link{set_motifs}()}
}
\concept{motif functions}
