% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getMotifById}
\alias{getMotifById}
\title{Retrieve motif objects by id}
\usage{
getMotifById(id)
}
\arguments{
\item{id}{character vector of motif identifiers}
}
\value{
A list of objects of class \code{RBPMotif}
}
\description{
Retrieves one or more motif objects identified by motif id.
}
\examples{
getMotifById("M178_0.6")

getMotifById(c("M178_0.6", "M188_0.6"))
}
\seealso{
Other motif functions: \code{\link{generateIUPACByKmers}},
  \code{\link{generateIUPACByMatrix}},
  \code{\link{generateKmersFromIUPAC}},
  \code{\link{getMotifByRBP}}, \code{\link{getMotifs}},
  \code{\link{getPPM}}, \code{\link{initIUPAClookupTable}},
  \code{\link{motifsMetaInfo}}, \code{\link{setMotifs}}
}
\concept{motif functions}
