% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getMotifByRBP}
\alias{getMotifByRBP}
\title{Retrieve motif objects by gene symbol}
\usage{
getMotifByRBP(rbp)
}
\arguments{
\item{rbp}{character vector of gene symbols of RNA-binding proteins}
}
\value{
A list of objects of class Motif
}
\description{
Retrieves one or more motif objects identified by gene symbol.
}
\examples{
getMotifByRBP("ELAVL1")

getMotifByRBP(c("ELAVL1", "ELAVL2"))
}
\seealso{
Other motif functions: \code{\link{generateIUPACByKmers}},
  \code{\link{generateIUPACByMatrix}},
  \code{\link{generateKmersFromIUPAC}},
  \code{\link{getMotifById}}, \code{\link{getMotifs}},
  \code{\link{getPPM}}, \code{\link{initIUPAClookupTable}},
  \code{\link{motifsMetaInfo}}, \code{\link{setMotifs}}
}
\concept{motif functions}
