% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getMotifs}
\alias{getMotifs}
\title{Retrieve list of all motifs}
\usage{
getMotifs()
}
\value{
A list of objects of class Motif
}
\description{
Retrieves all Transite motifs
}
\examples{
transite_motifs <- getMotifs()
}
\seealso{
Other motif functions: 
\code{\link{generateIUPACByKmers}()},
\code{\link{generateIUPACByMatrix}()},
\code{\link{generateKmersFromIUPAC}()},
\code{\link{getMotifById}()},
\code{\link{getMotifByRBP}()},
\code{\link{getPPM}()},
\code{\link{initIUPAClookupTable}()},
\code{\link{motifsMetaInfo}()},
\code{\link{setMotifs}()}
}
\concept{motif functions}
