% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getPPM}
\alias{getPPM}
\title{Get Position Probability Matrix (PPM) from motif object}
\usage{
getPPM(motif)
}
\arguments{
\item{motif}{object of class Motif}
}
\value{
The position probability matrix of the specified motif
}
\description{
Return the position probability matrix of the specified motif.
}
\examples{
getPPM(getMotifById('M178_0.6')[[1]])
}
\seealso{
Other motif functions: \code{\link{generateIUPACByKmers}},
  \code{\link{generateIUPACByMatrix}},
  \code{\link{generateKmersFromIUPAC}},
  \code{\link{getMotifById}}, \code{\link{getMotifByRBP}},
  \code{\link{getMotifs}},
  \code{\link{initIUPAClookupTable}},
  \code{\link{motifsMetaInfo}}, \code{\link{setMotifs}}
}
\concept{motif functions}
