% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{homopolymerCorrection}
\alias{homopolymerCorrection}
\title{Correction for Homopolymeric Stretches}
\usage{
homopolymerCorrection(sequences, k, kmers, is.rna = FALSE)
}
\arguments{
\item{sequences}{character vector of DNA or RNA sequences}

\item{k}{length of \emph{k}-mer, either \code{6} for hexamers or \code{7} for heptamers}

\item{kmers}{column sums of return value of
\code{Biostrings::oligonucleotideFrequency(sequences)}}

\item{is.rna}{if \code{sequences} are RNA sequences, this
flag needs to be set}
}
\value{
Returns a named numeric vector, where the elements are \emph{k}-mer counts and the
names are \emph{k}-mers.
}
\description{
Counts all non-overlapping instances of \emph{k}-mers in a given set of sequences.
}
\seealso{
Other \emph{k}-mer functions: \code{\link{calculateKmerEnrichment}},
  \code{\link{checkKmers}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{drawVolcanoPlot}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{\emph{k}-mer functions}
