% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{initIUPAClookupTable}
\alias{initIUPAClookupTable}
\title{Initializes the IUPAC lookup table}
\usage{
initIUPAClookupTable()
}
\value{
an environment, the IUPAC lookup hash table
}
\description{
Initializes a hash table that serves as a IUPAC lookup table for the
\code{\link{generateIUPACByMatrix}} function.
}
\details{
IUPAC RNA nucleotide code:
\tabular{rl}{
  \code{A} \tab Adenine\cr
  \code{C} \tab Cytosine\cr
  \code{G} \tab Guanine\cr
  \code{U} \tab Uracil\cr
  \code{R} \tab A or G\cr
  \code{Y} \tab C or U\cr
  \code{S} \tab G or C\cr
  \code{W} \tab A or U\cr
  \code{K} \tab G or U\cr
  \code{M} \tab A or C\cr
  \code{B} \tab C or G or U\cr
  \code{D} \tab A or G or U\cr
  \code{H} \tab A or C or U\cr
  \code{V} \tab A or C or G\cr
  \code{N} \tab any base
}
}
\examples{
generateIUPACByMatrix(motifMatrix(getMotifById("M178_0.6")[[1]]),
  code = initIUPAClookupTable())
}
\references{
\url{http://www.chem.qmul.ac.uk/iubmb/misc/naseq.html}
}
\seealso{
Other motif functions: 
\code{\link{generateIUPACByKmers}()},
\code{\link{generateIUPACByMatrix}()},
\code{\link{generateKmersFromIUPAC}()},
\code{\link{getMotifById}()},
\code{\link{getMotifByRBP}()},
\code{\link{getMotifs}()},
\code{\link{getPPM}()},
\code{\link{motifsMetaInfo}()},
\code{\link{setMotifs}()}
}
\concept{motif functions}
