% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kmers_enrichment}
\alias{kmers_enrichment}
\title{Example \emph{k}-mer Enrichment Data}
\format{
A data frame with the following columns:
\tabular{rl}{
  \code{kmer} \tab contains all hexamers (AAAAAA to UUUUUU)\cr
  \code{foreground_count} \tab absolute \emph{k}-mer frequency in
  foreground set\cr
  \code{background_count} \tab absolute \emph{k}-mer frequency in
  background set\cr
  \code{enrichment} \tab enrichment of \emph{k}-mer in foreground
  relative to background\cr
  \code{p_value} \tab associated p-value of enrichment\cr
  \code{adj_p_value} \tab multiple testing corrected p-value\cr
}
}
\usage{
kmers_enrichment
}
\description{
This data frame with \emph{k}-mer enrichment data (as produced by
\code{\link{runKmerTSMA}}) is used in a code example for k-mer volcano plot
function \code{\link{drawVolcanoPlot}}.
}
\keyword{datasets}
