% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookup_kmer_scores}
\alias{lookup_kmer_scores}
\title{\emph{k}-mer Score Lookup Table Access Function}
\usage{
lookup_kmer_scores(kmers, kmerScores)
}
\arguments{
\item{kmers}{list of \emph{k}-mers}

\item{kmerScores}{position weight matrix}
}
\value{
numeric vector of \emph{k}-mer scores
}
\description{
C++ implementation of \emph{k}-mer score hash table lookup.
}
