% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{motifs_meta_info}
\alias{motifs_meta_info}
\title{Displays motif meta information.}
\usage{
motifs_meta_info()
}
\value{
A data frame containing meta information for all Transite
motifs, with the
following columns:
\itemize{
  \item \code{id}
  \item \code{rbps}
  \item \code{length}
  \item \code{iupac}
  \item \code{type}
  \item \code{species}
  \item \code{src}
}
}
\description{
Generates a data frame with meta information about all Transite motifs.
}
\examples{
motifs_meta_info()
}
\seealso{
Other motif functions: 
\code{\link{generate_iupac_by_kmers}()},
\code{\link{generate_iupac_by_matrix}()},
\code{\link{generate_kmers_from_iupac}()},
\code{\link{get_motif_by_id}()},
\code{\link{get_motif_by_rbp}()},
\code{\link{get_motifs}()},
\code{\link{get_ppm}()},
\code{\link{init_iupac_lookup_table}()},
\code{\link{set_motifs}()}
}
\concept{motif functions}
