% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{motifsMetaInfo}
\alias{motifsMetaInfo}
\title{Displays motif meta information.}
\usage{
motifsMetaInfo()
}
\value{
A data frame containing meta information for all Transite
motifs, with the
following columns:
\itemize{
  \item \code{id}
  \item \code{rbps}
  \item \code{length}
  \item \code{iupac}
  \item \code{type}
  \item \code{species}
  \item \code{src}
}
}
\description{
Generates a data frame with meta information about all Transite motifs.
}
\examples{
motifsMetaInfo()
}
\seealso{
Other motif functions: 
\code{\link{generateIUPACByKmers}()},
\code{\link{generateIUPACByMatrix}()},
\code{\link{generateKmersFromIUPAC}()},
\code{\link{getMotifById}()},
\code{\link{getMotifByRBP}()},
\code{\link{getMotifs}()},
\code{\link{getPPM}()},
\code{\link{initIUPAClookupTable}()},
\code{\link{setMotifs}()}
}
\concept{motif functions}
