% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{permTestGeometricMean}
\alias{permTestGeometricMean}
\title{Permutation Test Based Significance of Observed Mean}
\usage{
permTestGeometricMean(actual.mean, motif.kmers, random.permutations,
  alternative = c("two.sided", "less", "greater"), conf.level = 0.95,
  produce.plot = TRUE)
}
\arguments{
\item{actual.mean}{observed mean}

\item{motif.kmers}{set of \emph{k}-mers that were used to compute the \code{actual.mean}}

\item{random.permutations}{a set of random permutations of the original data, used
to generate an empirical null distribution.}

\item{alternative}{side of the test, one of the following: \code{"two.sided"},
\code{"less"}, \code{"greater"}}

\item{conf.level}{confidence level for the returned confidence
    interval.}

\item{produce.plot}{if distribution plot should be part of the returned list}
}
\value{
A list with the following components:
\tabular{rl}{
  \code{p.value.estimate} \tab the estimated p-value of the observed mean\cr
  \code{conf.int} \tab the confidence interval around that estimate\cr
  \code{plot} \tab plot of the empirical distribution of geometric means of the
  enrichment values
}
}
\description{
\code{permTestGeometricMean} returns an estimate of the significance of the observed
mean, given a set of random permutations of the data.
}
\seealso{
Other \emph{k}-mer functions: \code{\link{calculateKmerEnrichment}},
  \code{\link{checkKmers}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{drawVolcanoPlot}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{\emph{k}-mer functions}
