% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{class}
\name{RBPMotif-class}
\alias{RBPMotif-class}
\alias{.RBPMotif}
\alias{motif.id,RBPMotif-method}
\alias{motif.id}
\alias{rbps,RBPMotif-method}
\alias{rbps}
\alias{motif.matrix,RBPMotif-method}
\alias{motif.matrix}
\alias{hexamers,RBPMotif-method}
\alias{hexamers}
\alias{heptamers,RBPMotif-method}
\alias{heptamers}
\alias{iupac,RBPMotif-method}
\alias{iupac}
\alias{motif.type,RBPMotif-method}
\alias{motif.type}
\alias{motif.species,RBPMotif-method}
\alias{motif.species}
\alias{motif.source,RBPMotif-method}
\alias{motif.source}
\alias{show,RBPMotif-method}
\alias{width,RBPMotif-method}
\alias{plot,RBPMotif,ANY-method}
\alias{plot,RBPMotif-method}
\title{An S4 class to represent a RBPMotif}
\usage{
motif.id(object)

\S4method{motif.id}{RBPMotif}(object)

rbps(object)

\S4method{rbps}{RBPMotif}(object)

motif.matrix(object)

\S4method{motif.matrix}{RBPMotif}(object)

hexamers(object)

\S4method{hexamers}{RBPMotif}(object)

heptamers(object)

\S4method{heptamers}{RBPMotif}(object)

iupac(object)

\S4method{iupac}{RBPMotif}(object)

motif.type(object)

\S4method{motif.type}{RBPMotif}(object)

motif.species(object)

\S4method{motif.species}{RBPMotif}(object)

motif.source(object)

\S4method{motif.source}{RBPMotif}(object)

\S4method{show}{RBPMotif}(object)

\S4method{width}{RBPMotif}(x)

\S4method{plot}{RBPMotif,ANY}(x)
}
\arguments{
\item{object}{RBPMotif object}

\item{x}{RBPMotif object}
}
\value{
Object of type RBPMotif
}
\description{
An S4 class to represent a RBPMotif

Getter Method motif.id

Getter Method rbps

Getter Method motif.matrix

Getter Method hexamers

Getter Method heptamers

Getter Method iupac

Getter Method motif.type

Getter Method motif.species

Getter Method motif.source
}
\section{Slots}{

\describe{
\item{\code{id}}{motif id (character vector of length 1)}

\item{\code{rbps}}{character vector of names of RNA-binding proteins associated
with this motif}

\item{\code{matrix}}{data frame with four columns (A, C, G, U) and 6 - 15 rows
(positions),
where cell (i, j) contains weight of nucleotide j on position i}

\item{\code{hexamers}}{character vector of hexamers associated with this motif}

\item{\code{heptamers}}{character vector of heptamers associated with this motif}

\item{\code{width}}{width of the motif (i.e., \code{nrow(matrix)})}

\item{\code{iupac}}{IUPAC code for motif matrix
(see \code{\link{generateIUPACByMatrix}})}

\item{\code{type}}{type of motif (e.g., \code{'HITS-CLIP'}, \code{'EMSA'},
\code{'SELEX'}, etc.)}

\item{\code{species}}{species where motif was discovered (e.g.,
\code{'Homo sapiens'})}

\item{\code{src}}{source of motif (e.g., \code{'RBPDB v1.3.1'})}
}}

\examples{
kmers <- c("AAAAAAA", "CAAAAAA")
iupac <- generateIUPACByKmers(kmers,
  code = initIUPAClookupTable())
hexamers <- generateKmersFromIUPAC(iupac, 6)
heptamers <- generateKmersFromIUPAC(iupac, 7)
new("RBPMotif", id = "custom.motif", rbps = "RBP1",
  matrix = NULL, hexamers = hexamers, heptamers = heptamers, width = 7L,
  iupac = iupac, type = "HITS-CLIP", species = "Homo sapiens", src = "user"
)
}
