% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runKmerSPMA}
\alias{runKmerSPMA}
\title{\emph{k}-mer-based Spectrum Motif Analysis}
\usage{
runKmerSPMA(background.set, motifs = NULL, k = 6, n.bins = 40,
  max.model.degree = 1, max.cs.permutations = 1e+07,
  min.cs.permutations = 5000, fg.permutations = 5000,
  p.adjust.method = "BH", p.combining.method = "fisher", n.cores = 1)
}
\arguments{
\item{background.set}{character vector of ranked sequences, either DNA
(only containing upper case characters A, C, G, T) or RNA (A, C, G, U).
The sequences in \code{background.set} must be ranked (i.e., sorted).
Commonly used sorting criteria are measures of differential expression, such
as fold change or signal-to-noise ratio (e.g., between treatment and control
samples in gene expression profiling experiments).}

\item{motifs}{a list of motifs that is used to score the specified sequences.
If \code{is.null(motifs)} then all Transite motifs are used.}

\item{k}{length of \emph{k}-mer, either \code{6} for hexamers or
\code{7} for heptamers}

\item{n.bins}{specifies the number of bins in which the sequences
will be divided,
valid values are between 7 and 100}

\item{max.model.degree}{maximum degree of polynomial}

\item{max.cs.permutations}{maximum number of permutations performed in
Monte Carlo test
for consistency score}

\item{min.cs.permutations}{minimum number of permutations performed in
Monte Carlo test
for consistency score}

\item{fg.permutations}{numer of foreground permutations}

\item{p.adjust.method}{see \code{\link[stats]{p.adjust}}}

\item{p.combining.method}{one of the following: Fisher (1932)
(\code{"fisher"}), Stouffer (1949),
Liptak (1958) (\code{"SL"}), Mudholkar and George (1979)
(\code{"MG"}), and Tippett (1931)
(\code{"tippett"}) (see \code{\link{pCombine}})}

\item{n.cores}{number of computing cores to use}
}
\value{
A list with the following components:
\tabular{rl}{
  \code{foreground.scores} \tab the result of \code{\link{runKmerTSMA}}
  for the binned data\cr
  \code{spectrum.info.df} \tab a data frame with the SPMA results\cr
  \code{spectrum.plots} \tab a list of spectrum plots, as generated by
  \code{\link{scoreSpectrum}}\cr
  \code{classifier.scores} \tab a list of classifier scores, as returned by
  \code{\link{spectrumClassifier}}
}
}
\description{
SPMA helps to illuminate the relationship between RBP binding evidence
and the transcript
sorting criterion, e.g., fold change between treatment and control samples.
}
\details{
In order to investigate how motif targets are distributed across a
spectrum of
transcripts (e.g., all transcripts of a platform, ordered by fold change),
Spectrum Motif Analysis visualizes the gradient of RBP binding evidence
across all transcripts.

The \emph{k}-mer-based approach differs from the matrix-based approach by
how the sequences are
scored. Here, sequences are broken into \emph{k}-mers, i.e.,
oligonucleotide sequences of
\emph{k} bases.
And only statistically significantly enriched or depleted \emph{k}-mers
are then used to
calculate a score for each RNA-binding protein, which quantifies its
target overrepresentation.
}
\examples{
# example data set
background.df <- transite:::ge$background
# sort sequences by signal-to-noise ratio
background.df <- dplyr::arrange(background.df, value)
# character vector of named and ranked (by signal-to-noise ratio) sequences
background.set <- gsub("T", "U", background.df$seq)
names(background.set) <- paste0(background.df$refseq, "|",
  background.df$seq.type)

results <- runKmerSPMA(background.set,
                       motifs = getMotifById("M178_0.6"),
                       n.bins = 20,
                       fg.permutations = 10)

\dontrun{
results <- runKmerSPMA(background.set)}

}
\seealso{
Other SPMA functions: \code{\link{runMatrixSPMA}},
  \code{\link{scoreSpectrum}},
  \code{\link{spectrumClassifier}},
  \code{\link{subdivideData}}

Other \emph{k}-mer functions: \code{\link{calculateKmerEnrichment}},
  \code{\link{checkKmers}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{drawVolcanoPlot}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerTSMA}}
}
\concept{SPMA functions}
\concept{\emph{k}-mer functions}
