% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runMatrixSPMA}
\alias{runMatrixSPMA}
\title{Matrix-based Spectrum Motif Analysis}
\usage{
runMatrixSPMA(background.set, motifs = NULL, n.bins = 40,
  max.model.degree = 1, max.cs.permutations = 1e+07,
  min.cs.permutations = 5000, max.hits = 5,
  threshold.method = "p.value", threshold.value = 0.25^6,
  max.fg.permutations = 1e+06, min.fg.permutations = 1000, e = 5,
  p.adjust.method = "BH", n.cores = 1, cache = paste0(tempdir(),
  "/sc/"))
}
\arguments{
\item{background.set}{a named character vector of background
sequences (naming follows same
rules as foreground set sequences)}

\item{motifs}{a list of motifs that is used to score the specified sequences.
If \code{is.null(motifs)} then all Transite motifs are used.}

\item{n.bins}{specifies the number of bins in which the sequences
will be divided,
valid values are between 7 and 100}

\item{max.model.degree}{maximum degree of polynomial}

\item{max.cs.permutations}{maximum number of permutations performed in
Monte Carlo test
for consistency score}

\item{min.cs.permutations}{minimum number of permutations performed in
Monte Carlo test
for consistency score}

\item{max.hits}{maximum number of putative binding sites per mRNA
that are counted}

\item{threshold.method}{either \code{"p.value"} (default) or
\code{"relative"}.
If \code{threshold.method} equals \code{"p.value"}, the default
\code{threshold.value}
 is \code{0.25^6}, which is
lowest p-value that can be achieved by hexamer motifs, the shortest
supported motifs.
If \code{threshold.method} equals \code{"relative"}, the default
\code{threshold.value}
is \code{0.9}, which is 90\% of the maximum PWM score.}

\item{threshold.value}{semantics of the \code{threshold.value} depend on
\code{threshold.method} (default is 0.25^6)}

\item{max.fg.permutations}{maximum number of foreground permutations
performed in
Monte Carlo test for enrichment score}

\item{min.fg.permutations}{minimum number of foreground permutations
performed in
Monte Carlo test for enrichment score}

\item{e}{integer-valued stop criterion for enrichment score Monte Carlo test: aborting
permutation process after
observing \code{e} random enrichment values with more extreme values than
the actual
enrichment value}

\item{p.adjust.method}{adjustment of p-values from Monte Carlo tests to
avoid alpha error
 accumulation, see \code{\link[stats]{p.adjust}}}

\item{n.cores}{the number of cores that are used}

\item{cache}{either logical or path to a directory where scores are cached.
The scores of each
motif are stored in a
separate file that contains a hash table with RefSeq identifiers and
sequence type
qualifiers as keys and the number of putative binding sites as values.
If \code{cache} is \code{FALSE}, scores will not be cached.}
}
\value{
A list with the following components:
\tabular{rl}{
  \code{foreground.scores} \tab the result of \code{\link{scoreTranscripts}}
  for the foreground
  sets (the bins)\cr
  \code{background.scores} \tab the result of \code{\link{scoreTranscripts}}
  for the background
  set\cr
  \code{enrichment.dfs} \tab a list of data frames, returned by
  \code{\link{calculateMotifEnrichment}}\cr
  \code{spectrum.info.df} \tab a data frame with the SPMA results\cr
  \code{spectrum.plots} \tab a list of spectrum plots, as generated by
  \code{\link{scoreSpectrum}}\cr
  \code{classifier.scores} \tab a list of classifier scores, as returned by
  \code{\link{spectrumClassifier}}
}
}
\description{
SPMA helps to illuminate the relationship between RBP binding
evidence and the transcript
sorting criterion, e.g., fold change between treatment and control samples.
}
\details{
In order to investigate how motif targets are distributed across a
spectrum of
transcripts (e.g., all transcripts of a platform, ordered by fold change),
Spectrum Motif Analysis visualizes the gradient of RBP binding evidence
across all transcripts.

The matrix-based approach skips the \emph{k}-merization step of the
\emph{k}-mer-based approach
and instead scores the transcript sequence as a whole with a position
specific scoring matrix.

For each sequence in foreground and background sets and each sequence motif,
the scoring algorithm evaluates the score for each sequence position.
Positions with
a relative score greater than a certain threshold are considered hits, i.e.,
putative binding sites.

By scoring all sequences in foreground and background sets, a hit count
for each motif and
each set is obtained, which is used to calculate enrichment values and
associated p-values
in the same way in which motif-compatible hexamer enrichment values are
calculated in
the \emph{k}-mer-based approach. P-values are adjusted with one of the
available adjustment methods.

An advantage of the matrix-based approach is the possibility of detecting
clusters of
binding sites. This can be done by counting regions with many hits using
positional
hit information or by simply applying a hit count threshold per
sequence, e.g., only
sequences with more than some number of hits are considered. Homotypic
clusters of RBP
binding sites may play a similar role as clusters of transcription factors.
}
\examples{
\dontrun{
# example data set
background.df <- transite:::ge$background
# sort sequences by signal-to-noise ratio
background.df <- dplyr::arrange(background.df, value)
# character vector of named sequences
background.set <- gsub("T", "U", background.df$seq)
names(background.set) <- paste0(background.df$refseq, "|",
  background.df$seq.type)

results <- runMatrixSPMA(background.set)
}

}
\seealso{
Other SPMA functions: \code{\link{runKmerSPMA}},
  \code{\link{scoreSpectrum}},
  \code{\link{spectrumClassifier}},
  \code{\link{subdivideData}}

Other matrix functions: \code{\link{calculateMotifEnrichment}},
  \code{\link{runMatrixTSMA}},
  \code{\link{scoreTranscriptsSingleMotif}},
  \code{\link{scoreTranscripts}}
}
\concept{SPMA functions}
\concept{matrix functions}
