% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreSequences}
\alias{scoreSequences}
\title{Score Sequences with PWM}
\usage{
scoreSequences(sequences, pwm)
}
\arguments{
\item{sequences}{list of sequences}

\item{pwm}{position weight matrix}
}
\value{
list of PWM scores for each sequence
}
\description{
C++ implementation of PWM scoring algorithm
}
\examples{
motif <- getMotifById("M178_0.6")[[1]]
sequences <- c("CAACAGCCTTAATT", "CAGTCAAGACTCC", "CTTTGGGGAAT",
               "TCATTTTATTAAA", "AATTGGTGTCTGGATACTTCCCTGTACAT",
               "ATCAAATTA", "TGTGGGG", "GACACTTAAAGATCCT",
               "TAGCATTAACTTAATG", "ATGGA", "GAAGAGTGCTCA", "ATAGAC",
               "AGTTC", "CCAGTAA")
seq.char.vectors <- lapply(sequences, function(seq) {
  unlist(strsplit(seq, ""))
})
scoreSequences(seq.char.vectors, as.matrix(motif$matrix))

}
