% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{scoreSpectrum}
\alias{scoreSpectrum}
\title{Calculates spectrum scores and creates spectrum plots}
\usage{
scoreSpectrum(x, p.value = array(1, length(x)),
  x.label = "log enrichment", midpoint = 0, max.model.degree = 3,
  max.cs.permutations = 1e+07, min.cs.permutations = 5000, e = 5)
}
\arguments{
\item{x}{vector of values (e.g., enrichment values, normalized RBP
scores) per bin}

\item{p.value}{vector of p-values (e.g., significance of enrichment
values) per bin}

\item{x.label}{label of values (e.g., \code{"enrichment value"})}

\item{midpoint}{for enrichment values the midpoint should be \code{1},
for log enrichment
values \code{0})}

\item{max.model.degree}{maximum degree of polynomial}

\item{max.cs.permutations}{maximum number of permutations performed in
Monte Carlo test
for consistency score}

\item{min.cs.permutations}{minimum number of permutations performed in
Monte Carlo test
for consistency score}

\item{e}{stop criterion for consistency score Monte Carlo test: aborting
permutation
process after
observing \code{e} random consistency values with more extreme values than
the actual
consistency value}
}
\value{
A list object of class \code{SpectrumScore} with the following
components:
\tabular{rl}{
  \code{adj.r.squared} \tab adjusted \eqn{R^2} of polynomial model\cr
  \code{degree} \tab maximum degree of polynomial\cr
  \code{residuals} \tab residuals  of polynomial model\cr
  \code{slope} \tab coefficient of the linear term of the polynomial model
   (spectrum "direction")\cr
  \code{f.statistic} \tab statistic of the F-test \cr
  \code{f.statistic.p.value} \tab p-value of F-test\cr
  \code{consistency.score} \tab normalized sum of deviance between the linear
  interpolation of the scores
   of two adjoining bins and the score of the middle bin, for each position
   in the spectrum \cr
  \code{consistency.score.p.value} \tab obtained by Monte Carlo sampling
  (randomly permuting
  the coordinates of the scores vector) \cr
  \code{consistency.score.n} \tab number of permutations\cr
  \code{plot} \tab
}
}
\description{
Spectrum scores are a means to evaluate if a spectrum has a meaningful
(i.e., biologically relevant) or a random pattern.
}
\details{
One way to quantify the meaningfulness of a spectrum is to calculate the
deviance
between the linear interpolation of the scores of two adjoining bins and
the score
of the middle bin, for each position in the spectrum. The lower the score,
the more consistent
the trend in the spectrum plot. Formally, the local consistency score
\eqn{x_c} is defined as
\deqn{x_c = \frac{1}{n} \sum_{i = 1}^{n - 2}{\left|\frac{s_i + s_{i + 2}}{2} - s_{i + 1}\right|}.}
In order to obtain an estimate of the significance of a particular
score \eqn{x_c'},
Monte Carlo sampling
is performed by randomly permuting the coordinates of the scores
vector \eqn{s}
and recomputing \eqn{x_c}.
The probability estimate \eqn{\hat{p}} is given by the lower tail
version of the cumulative
distribution function
\deqn{\hat{Pr}(T(x)) &= \frac{\sum\limits_{i = 1}^n \mathbbm{1}\left(T(y_i) \le T(x)\right) + 1}{n + 1},}
where \eqn{\mathbbm{1}} is the indicator function, \eqn{n} is the
sample size, i.e.,
the number of performed permutations, and \eqn{T} equals \eqn{x_c}
in the above equation.

An alternative approach to assess the consistency of a spectrum plot is
via polynomial regression.
In a first step, polynomial regression models of various degrees are
fitted to the data, i.e.,
the dependent variable \eqn{\boldsymbol{s}} (vector of scores), and
orthogonal
polynomials of the independent variable \eqn{\boldsymbol{b}} (vector of
bin numbers).
Secondly, the model that reflects best the true nature of the data is
selected by
means of the F-test. And lastly, the adjusted \eqn{R^2} and the sum of
squared residuals
are calculated to indicate how well the model fits the data. These
statistics are used as
scores to rank the spectrum plots.
In general, the polynomial regression equation is
\deqn{y_i = \beta_0 + \beta_1 x_i + \beta_2 x_i^2 + \cdots + \beta_m x_i^m + \epsilon_i,}
where \eqn{m} is the degree of the polynomial (usually \eqn{m \le 5}),
and \eqn{\epsilon_i}
is the error term.
The dependent variable \eqn{\boldsymbol{y}} is the vector of
scores \eqn{\boldsymbol{s}}
and \eqn{\boldsymbol{x}}
to \eqn{\boldsymbol{x}^m} are the orthogonal polynomials of
the vector of bin
numbers \eqn{\boldsymbol{b}}.
Orthogonal polynomials are used in order to reduce the correlation
between the different
powers of \eqn{\boldsymbol{b}}
and therefore avoid multicollinearity in the model. This is important,
because correlated
predictors lead to unstable coefficients, i.e., the coefficients of a
polynomial
regression model of
degree \eqn{m} can be greatly different from a model of degree \eqn{m + 1}.

The orthogonal polynomials of vector \eqn{\boldsymbol{b}} are obtained by
centering (subtracting the mean),
QR decomposition, and subsequent normalization.
Given the dependent variable \eqn{\boldsymbol{y}} and the orthogonal
polynomials
of \eqn{\boldsymbol{b}} \eqn{\boldsymbol{x}} to
\eqn{\boldsymbol{x}^m}, the model coefficients \eqn{\boldsymbol{\beta}}
are chosen
in a way to minimize
the deviance between the actual and the predicted values characterized by
\deqn{\mathcal{M}(x) = \beta_0 + \beta_1 x + \beta_2 x^2 + \cdots + \beta_m x^m}
\deqn{\mathcal{M} = \argmin_{\mathcal{M}}{\left(\sum_{i = 1}^n{L(y_i, \mathcal{M}(x_i))}\right)},}
where \eqn{L(\text{actual value}, \text{predicted value})} denotes
the loss function.

Ordinary least squares is used as estimation method for the model
coefficients \eqn{\boldsymbol{\beta}}. The loss
function of ordinary least squares is the sum of squared residuals
(\eqn{\mathit{SSR}})
and is defined as follows
\deqn{\mathit{SSR}(\boldsymbol{y},
\boldsymbol{\hat{y}}) = \sum_{i = 1}^n{(y_i - \hat{y}_i)^2},}
where \eqn{\boldsymbol{y}} are the observed data
and \eqn{\boldsymbol{\hat{y}}} the
model predictions.

Thus the ordinary least squares estimate of the
coefficients \eqn{\boldsymbol{\hat{\beta}}}
(including the
intercept \eqn{\hat{\beta}_0}) of the model \eqn{\mathcal{M}} is defined by
\deqn{\boldsymbol{\hat{\beta}} = \argmin_{\boldsymbol{\beta}}{\left(\sum_{i = 1}^n{\left(y_i - \beta_0 - \sum_{j = 1}^m{\beta_j x_i^j}\right)^2}\right)}.}

After polynomial models of various degrees have been fitted to the
data, the F-test
is used to select
the model that best fits the data. Since the \eqn{\mathit{SSR}}
monotonically decreases with
increasing model degree (model complexity), the relative decrease
of the \eqn{\mathit{SSR}}
between the
simpler model and the more complex model must outweigh the increase
in model complexity
between the two
models. The F-test gives the probability that a relative decrease of
the \eqn{\mathit{SSR}}
between the simpler
and the more complex model given their respective degrees of freedom is due
to chance.
A low p-value
indicates that the additional degrees of freedom of the more complex model
lead to a
better fit of
the data than would be expected after a mere increase of degrees of freedom.

The F-statistic is calculated as follows
\deqn{F = \frac{(\mathit{SSR}_1 - \mathit{SSR}_2) / (p_2 - p_1)}{\mathit{SSR}_2 / (n - p_2)},}
where \eqn{\mathit{SSR}_i} is the sum of squared residuals
and \eqn{p_i} is the number
of parameters of
model \eqn{i}. The number of data points, i.e., bins, is denoted as \eqn{n}.
\eqn{F} is distributed according to the F-distribution
with \eqn{df_1 = p_2 - p_1} and \eqn{df_2 = n - p_2}.
}
\examples{
# random spectrum
scoreSpectrum(runif(n = 40, min = -1, max = 1), max.model.degree = 1)

# non-random linear spectrum
signal <- seq(-1, 0.99, 2 / 40)
noise <- rnorm(n = 40, mean = 0, sd = 0.5)
scoreSpectrum(signal + noise, max.model.degree = 1,
  max.cs.permutations = 100000)

# non-random quadratic spectrum
signal <- seq(-1, 0.99, 2 / 40)^2 - 0.5
noise <- rnorm(n = 40, mean = 0, sd = 0.2)
scoreSpectrum(signal + noise, max.model.degree = 2,
  max.cs.permutations = 100000)
}
\seealso{
Other SPMA functions: \code{\link{runKmerSPMA}},
  \code{\link{runMatrixSPMA}},
  \code{\link{spectrumClassifier}},
  \code{\link{subdivideData}}
}
\concept{SPMA functions}
