% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setMotifs}
\alias{setMotifs}
\title{Set Transite motif database}
\usage{
setMotifs(value)
}
\arguments{
\item{value}{list of Motif objects}
}
\description{
Globally sets Transite motif database, use with care.
}
\seealso{
Other motif functions: \code{\link{generateIUPACByKmers}},
  \code{\link{generateIUPACByMatrix}},
  \code{\link{generateKmersFromIUPAC}},
  \code{\link{getMotifById}}, \code{\link{getMotifByRBP}},
  \code{\link{getMotifs}}, \code{\link{getPPM}},
  \code{\link{initIUPAClookupTable}},
  \code{\link{motifsMetaInfo}}
}
\concept{motif functions}
