% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy.motif.matrix}
\alias{toy.motif.matrix}
\title{A toy motif matrix.}
\format{A data frame with four columns (A, C, G, U) and seven rows (position 1 - 7)}
\usage{
toy.motif.matrix
}
\description{
A toy motif matrix.
}
\keyword{datasets}
