% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_motif_matrix}
\alias{toy_motif_matrix}
\title{Toy Motif Matrix}
\format{
A data frame with four columns (A, C, G, U) and seven
rows (position 1 - 7)
}
\usage{
toy_motif_matrix
}
\description{
This toy motif matrix is used in code examples for various functions.
}
\keyword{datasets}
