% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transite-package.R
\docType{package}
\name{transite}
\alias{transite-package}
\alias{transite}
\title{transite}
\description{
transite is a computational method that allows comprehensive analysis of
the regulatory
role of RNA-binding proteins in various cellular processes by leveraging
preexisting
gene expression data and current knowledge of binding preferences of
}
\seealso{
Useful links:
\itemize{
  \item \url{https://transite.mit.edu}
}

}
\author{
Konstantin Krismer
}
\keyword{internal}
