% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{calculateKmerEnrichment}
\alias{calculateKmerEnrichment}
\title{\emph{k}-mer Enrichment between Foreground and Background Sets}
\usage{
calculateKmerEnrichment(foreground.sets, background.set, k,
  permutation = FALSE, chisq.p.value.threshold = 0.05,
  p.adjust.method = "BH", n.cores = 4)
}
\arguments{
\item{foreground.sets}{list of foreground sets; a foreground set is a
character vector of
DNA or RNA sequences (not both) and a strict subset of the
\code{background.set}}

\item{background.set}{character vector of DNA or RNA sequences that
constitute the
background set}

\item{k}{length of \emph{k}-mer, either \code{6} for hexamers or
\code{7} for heptamers}

\item{permutation}{if \code{TRUE}, only the enrichment value is returned
(efficiency mode
used for permutation testing)}

\item{chisq.p.value.threshold}{threshold below which Fisher's exact test
is used instead of Pearson's chi-squared test}

\item{p.adjust.method}{see \code{\link[stats]{p.adjust}}}

\item{n.cores}{number of computing cores to use}
}
\value{
A list with two entries:

(1) dfs: a list of data frames with results from
\code{\link{computeKmerEnrichment}} for each of the foreground sets
(2) kmers: a character vector of all k-mers
}
\description{
Calls \code{\link{computeKmerEnrichment}} to compute \emph{k}-mer
enrichment values
for multiple foregrounds. Calculates enrichment for foreground sets in
parallel.
}
\examples{
# define simple sequence sets for foreground and background
foreground.set1 <- c(
  "CAACAGCCUUAAUU", "CAGUCAAGACUCC", "CUUUGGGGAAU",
  "UCAUUUUAUUAAA", "AAUUGGUGUCUGGAUACUUCCCUGUACAU",
  "AUCAAAUUA", "AGAU", "GACACUUAAAGAUCCU",
  "UAGCAUUAACUUAAUG", "AUGGA", "GAAGAGUGCUCA",
  "AUAGAC", "AGUUC", "CCAGUAA"
)
foreground.set2 <- c("UUAUUUA", "AUCCUUUACA", "UUUUUUU", "UUUCAUCAUU")
foreground.sets <- list(foreground.set1, foreground.set2)
background.set <- c(foreground.set1, foreground.set2,
                    "CCACACAC", "CUCAUUGGAG", "ACUUUGGGACA", "CAGGUCAGCA")

# single-threaded
kmer.enrichment.values.st <- calculateKmerEnrichment(foreground.sets,
  background.set, 6, n.cores = 1)

# multi-threaded
kmer.enrichment.values.mt <- calculateKmerEnrichment(foreground.sets,
  background.set, 6)
}
\seealso{
Other \emph{k}-mer functions: \code{\link{checkKmers}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{drawVolcanoPlot}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{\emph{k}-mer functions}
