% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{calculateKmerEnrichment}
\alias{calculateKmerEnrichment}
\title{Correction for homopolymeric stretches}
\usage{
calculateKmerEnrichment(foreground.sets, background.set, k,
  permutation = FALSE, chisq.p.value.threshold = 0.05,
  p.adjust.method = "BH", n.cores = 4)
}
\arguments{
\item{foreground.sets}{list of foreground sets; a foreground set is a character vector of
DNA or RNA sequences (not both) and a strict subset of the \code{background.set}}

\item{background.set}{character vector of DNA or RNA sequences that constitute the
background set}

\item{k}{length of \emph{k}-mer, either \code{6} for hexamers or \code{7} for heptamers}

\item{permutation}{if \code{TRUE}, only the enrichment value is returned (efficiency mode
used for permutation testing)}

\item{chisq.p.value.threshold}{threshold below which Fisher's exact test is used instead of
\eqn{\Chi^2} test}

\item{p.adjust.method}{see \code{\link[stats]{p.adjust}}}

\item{n.cores}{number of computing cores to use}
}
\description{
Counts all non-overlapping instances of \emph{k}-mers in a given set of sequences.
}
\seealso{
Other \emph{k}-mer functions: \code{\link{checkKmers}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{drawVolcanoPlot}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{\emph{k}-mer functions}
