% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateKmerScores}
\alias{calculateKmerScores}
\title{\emph{k}-mer Score Calculation}
\usage{
calculateKmerScores(kmers, pwm)
}
\arguments{
\item{kmers}{list of \emph{k}-mers}

\item{pwm}{position weight matrix}
}
\value{
list of PWM scores for the specified \emph{k}-mers
}
\description{
C++ implementation of \emph{k}-mer score calculation
}
