% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateLocalConsistency}
\alias{calculateLocalConsistency}
\title{Local Consistency Score}
\usage{
calculateLocalConsistency(x, numPermutations, minPermutations, e)
}
\arguments{
\item{x}{numeric vector that contains values for shuffling}

\item{numPermutations}{maximum number of permutations performed in
Monte Carlo test
for consistency score}

\item{minPermutations}{minimum number of permutations performed in
Monte Carlo test
for consistency score}

\item{e}{stop criterion for consistency score Monte Carlo test:
aborting permutation
process after observing \code{e} random consistency values with
more extreme values
than the actual consistency value}
}
\value{
list with \code{score}, \code{p.value}, and \code{n} components
}
\description{
C++ implementation of Local Consistency Score algorithm.
}
\examples{
poor.enrichment.spectrum <- c(0.1, 0.5, 0.6, 0.4,
  0.7, 0.6, 1.2, 1.1, 1.8, 1.6)
local.consistency <- calculateLocalConsistency(poor.enrichment.spectrum,
  1000000, 1000, 5)

enrichment.spectrum <- c(0.1, 0.3, 0.6, 0.7, 0.8,
  0.9, 1.2, 1.4, 1.6, 1.4)
local.consistency <- calculateLocalConsistency(enrichment.spectrum,
  1000000, 1000, 5)
}
