% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateTranscriptMC}
\alias{calculateTranscriptMC}
\title{Local Consistency Score}
\usage{
calculateTranscriptMC(absoluteHits, totalSites, relHitsForeground, n,
  maxPermutations, minPermutations, e)
}
\arguments{
\item{absoluteHits}{number of putative binding sites per sequence
(returned by \code{\link{scoreTranscripts}})}

\item{totalSites}{number of potential binding sites per sequence
(returned by \code{\link{scoreTranscripts}})}

\item{relHitsForeground}{relative number of hits in foreground set}

\item{n}{number of sequences in the foreground set}

\item{maxPermutations}{maximum number of foreground permutations performed in
Monte Carlo test for enrichment score}

\item{minPermutations}{minimum number of foreground permutations performed in
Monte Carlo test for enrichment score}

\item{e}{stop criterion for enrichment score Monte Carlo test: aborting permutation process
after observing \code{e} random enrichment values with more extreme values than the actual
enrichment value}
}
\description{
C++ implementation of Local Consistency Score algorithm.
}
