% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateTranscriptMC}
\alias{calculateTranscriptMC}
\title{Motif Enrichment calculation}
\usage{
calculateTranscriptMC(absoluteHits, totalSites, relHitsForeground, n,
  maxPermutations, minPermutations, e)
}
\arguments{
\item{absoluteHits}{number of putative binding sites per sequence
(returned by \code{\link{scoreTranscripts}})}

\item{totalSites}{number of potential binding sites per sequence
(returned by \code{\link{scoreTranscripts}})}

\item{relHitsForeground}{relative number of hits in foreground set}

\item{n}{number of sequences in the foreground set}

\item{maxPermutations}{maximum number of foreground permutations performed in
Monte Carlo test for enrichment score}

\item{minPermutations}{minimum number of foreground permutations performed in
Monte Carlo test for enrichment score}

\item{e}{stop criterion for enrichment score Monte Carlo test: aborting permutation process
after observing \code{e} random enrichment values with more extreme values than the actual
enrichment value}
}
\value{
list with p-value and number of iterations of Monte Carlo sampling
for local consistency score
}
\description{
C++ implementation of Motif Enrichment calculation
}
\examples{
foreground.seqs <- c("CAGTCAAGACTCC", "AATTGGTGTCTGGATACTTCCCTGTACAT", "AGAT", "CCAGTAA")
background.seqs <- c("CAACAGCCTTAATT", "CAGTCAAGACTCC", "CTTTGGGGAAT",
                     "TCATTTTATTAAA", "AATTGGTGTCTGGATACTTCCCTGTACAT",
                     "ATCAAATTA", "AGAT", "GACACTTAAAGATCCT",
                     "TAGCATTAACTTAATG", "ATGGA", "GAAGAGTGCTCA",
                     "ATAGAC", "AGTTC", "CCAGTAA")
foreground.scores <- scoreTranscripts(foreground.seqs, cache = FALSE)
background.scores <- scoreTranscripts(background.seqs, cache = FALSE)

fg <- dplyr::filter(foreground.scores$df, motif.id == "M178_0.6")
bg <- dplyr::filter(background.scores$df, motif.id == "M178_0.6")

mc.result <- calculateTranscriptMC(bg$absolute.hits, bg$total.sites,
                                   fg$absolute.hits / fg$total.sites,
                                   length(foreground.seqs), 10000, 5000, 5)

}
