% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{computeKmerEnrichment}
\alias{computeKmerEnrichment}
\title{\emph{k}-mer Enrichment between Foreground and Background Sets}
\usage{
computeKmerEnrichment(foreground.kmers, background.kmers,
  permutation = FALSE, chisq.p.value.threshold = 0.05,
  p.adjust.method = "BH")
}
\arguments{
\item{foreground.kmers}{\emph{k}-mer counts of the foreground set
(generated by \code{\link{generateKmers}})}

\item{background.kmers}{\emph{k}-mer counts of the background set
(generated by \code{\link{generateKmers}})}

\item{permutation}{if \code{TRUE}, only the enrichment value is returned (efficiency mode
used for permutation testing)}

\item{chisq.p.value.threshold}{threshold below which Fisher's exact test is used instead of
\eqn{\Chi^2} test}

\item{p.adjust.method}{see \code{\link[stats]{p.adjust}}}
}
\value{
enrichment of \emph{k}-mers in specified foreground sequences.
A data frame with the following columns is returned:
\tabular{rl}{
  \code{foreground.count} \tab foreground counts for each \emph{k}-mer\cr
  \code{background.count} \tab background counts for each \emph{k}-mer\cr
  \code{enrichment} \tab \emph{k}-mer enrichment\cr
  \code{p.value} \tab p-value of \emph{k}-mer enrichment (either from Fisher's exact test or
  \eqn{\Chi^2} test.\cr
  \code{adj.p.value} \tab multiple testing corrected p-value\cr
}
}
\description{
Compares foreground sequence set to background sequence set and computes enrichment values
for each possible \emph{k}-mer.
}
\details{
Usually uses Pearson's \eqn{\Chi^2} test, but recalculates p-values with Fisher's exact test
for \eqn{\Chi^2} test p-values \code{<= chisq.p.value.threshold}. The reason this is done is
computational efficiency. Fisher's exact tests are computationally demanding and are only
performed in situations, where exact p-values are preferred, e.g., if expected < 5 or
significant p-values.
}
\seealso{
Other \emph{k}-mer functions: \code{\link{calculateKmerEnrichment}},
  \code{\link{checkKmers}}, \code{\link{drawVolcanoPlot}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{\emph{k}-mer functions}
