% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeMotifScore}
\alias{computeMotifScore}
\title{Motif Score Algorithm}
\usage{
computeMotifScore(kmers)
}
\arguments{
\item{kmers}{list of \emph{k}-mers}
}
\value{
data frame with columns \code{score}, \code{top_kmer},
and \code{top_kmer_enrichment}
}
\description{
C++ implementation of motif score algorithm.
}
