% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{drawVolcanoPlot}
\alias{drawVolcanoPlot}
\title{\emph{k}-mer Enrichment Volcano Plot}
\usage{
drawVolcanoPlot(kmers, motif.kmers, motif.rbps,
  significance.threshold = 0.01, show.legend = TRUE)
}
\arguments{
\item{kmers}{data frame with the following columns: kmer,
adj.p.value, enrichment}

\item{motif.kmers}{set of \emph{k}-mers that are associated with
a certain motif,
will be highlighted in volcano plot}

\item{motif.rbps}{name of RNA-binding proteins associated with
highlighted \emph{k}-mers
(character vector of length 1)}

\item{significance.threshold}{p-value threshold for significance,
e.g., \code{0.05} or
\code{0.01}}

\item{show.legend}{whether or not a legend should be shown}
}
\value{
volcano plot
}
\description{
Uses a volcano plot to visualize \emph{k}-mer enrichment.
X-axis is \eqn{\log_2} enrichment value,
y-axis is \eqn{\log_10} significance, i.e., multiple testing
corrected p-value from
Fisher's exact test or Pearson's chi-squared test.
}
\seealso{
Other TSMA functions: \code{\link{runKmerTSMA}},
  \code{\link{runMatrixTSMA}}

Other \emph{k}-mer functions: \code{\link{calculateKmerEnrichment}},
  \code{\link{checkKmers}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{empiricalEnrichmentMeanCDF}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{TSMA functions}
\concept{\emph{k}-mer functions}
