% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer-based.R
\name{empiricalEnrichmentMeanCDF}
\alias{empiricalEnrichmentMeanCDF}
\title{Significance of Observed Mean}
\usage{
empiricalEnrichmentMeanCDF(random.means, actual.mean,
  alternative = c("two.sided", "less", "greater"), conf.level = 0.95)
}
\arguments{
\item{random.means}{numeric vector of means based on random permutations of the data
(empirical null distribution)}

\item{actual.mean}{observed mean}

\item{alternative}{side of the test, one of the following: \code{"two.sided"},
\code{"less"}, \code{"greater"}}

\item{conf.level}{confidence level for the returned confidence
    interval.}
}
\value{
A list with the following components:
\tabular{rl}{
  \code{p.value.estimate} \tab the estimated p-value of the observed mean\cr
  \code{conf.int} \tab the confidence interval around that estimate
}
}
\description{
\code{empiricalEnrichmentMeanCDF} returns an estimate of the significance of the observed
mean, given a vector of means based on random permutations of the data.
}
\seealso{
Other \emph{k}-mer functions: \code{\link{calculateKmerEnrichment}},
  \code{\link{checkKmers}},
  \code{\link{computeKmerEnrichment}},
  \code{\link{drawVolcanoPlot}},
  \code{\link{generateKmers}},
  \code{\link{generatePermutedEnrichments}},
  \code{\link{homopolymerCorrection}},
  \code{\link{permTestGeometricMean}},
  \code{\link{runKmerSPMA}}, \code{\link{runKmerTSMA}}
}
\concept{\emph{k}-mer functions}
