% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{class}
\name{RBPMotif-class}
\alias{RBPMotif-class}
\alias{.RBPMotif}
\alias{motifId,RBPMotif-method}
\alias{motifId}
\alias{motifRbps,RBPMotif-method}
\alias{motifRbps}
\alias{motifMatrix,RBPMotif-method}
\alias{motifMatrix}
\alias{motifHexamers,RBPMotif-method}
\alias{motifHexamers}
\alias{motifHeptamers,RBPMotif-method}
\alias{motifHeptamers}
\alias{motifLength,RBPMotif-method}
\alias{motifLength}
\alias{motifIUPAC,RBPMotif-method}
\alias{motifIUPAC}
\alias{motifType,RBPMotif-method}
\alias{motifType}
\alias{motifSpecies,RBPMotif-method}
\alias{motifSpecies}
\alias{motifSource,RBPMotif-method}
\alias{motifSource}
\alias{show,RBPMotif-method}
\alias{plot,RBPMotif,ANY-method}
\alias{plot,RBPMotif-method}
\title{An S4 class to represent a RBPMotif}
\usage{
motifId(object)

\S4method{motifId}{RBPMotif}(object)

motifRbps(object)

\S4method{motifRbps}{RBPMotif}(object)

motifMatrix(object)

\S4method{motifMatrix}{RBPMotif}(object)

motifHexamers(object)

\S4method{motifHexamers}{RBPMotif}(object)

motifHeptamers(object)

\S4method{motifHeptamers}{RBPMotif}(object)

motifLength(object)

\S4method{motifLength}{RBPMotif}(object)

motifIUPAC(object)

\S4method{motifIUPAC}{RBPMotif}(object)

motifType(object)

\S4method{motifType}{RBPMotif}(object)

motifSpecies(object)

\S4method{motifSpecies}{RBPMotif}(object)

motifSource(object)

\S4method{motifSource}{RBPMotif}(object)

\S4method{show}{RBPMotif}(object)

\S4method{plot}{RBPMotif,ANY}(x)
}
\arguments{
\item{object}{RBPMotif object}

\item{x}{RBPMotif object}
}
\value{
Object of type RBPMotif
}
\description{
An S4 class to represent a RBPMotif

Getter Method motifId

Getter Method motifRbps

Getter Method motifMatrix

Getter Method motifHexamers

Getter Method motifHeptamers

Getter Method motifLength

Getter Method motifIUPAC

Getter Method motifType

Getter Method motifSpecies

Getter Method motifSource
}
\section{Slots}{

\describe{
\item{\code{id}}{motif id (character vector of length 1)}

\item{\code{rbps}}{character vector of names of RNA-binding proteins associated
with this motif}

\item{\code{matrix}}{data frame with four columns (A, C, G, U) and 6 - 15 rows
(positions),
where cell (i, j) contains weight of nucleotide j on position i}

\item{\code{hexamers}}{character vector of hexamers associated with this motif}

\item{\code{heptamers}}{character vector of heptamers associated with this motif}

\item{\code{length}}{length of the motif (i.e., \code{nrow(matrix)})}

\item{\code{iupac}}{IUPAC code for motif matrix
(see \code{\link{generate_iupac_by_matrix}})}

\item{\code{type}}{type of motif (e.g., \code{'HITS-CLIP'}, \code{'EMSA'},
\code{'SELEX'}, etc.)}

\item{\code{species}}{species where motif was discovered (e.g.,
\code{'Homo sapiens'})}

\item{\code{src}}{source of motif (e.g., \code{'RBPDB v1.3.1'})}
}}

\examples{
kmers <- c("AAAAAAA", "CAAAAAA")
iupac <- generate_iupac_by_kmers(kmers,
  code = init_iupac_lookup_table())
hexamers <- generate_kmers_from_iupac(iupac, 6)
heptamers <- generate_kmers_from_iupac(iupac, 7)
new("RBPMotif", id = "custom_motif", rbps = "RBP1",
  matrix = NULL, hexamers = hexamers, heptamers = heptamers, length = 7L,
  iupac = iupac, type = "HITS-CLIP", species = "Homo sapiens", src = "user"
)
}
