% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreSequences}
\alias{scoreSequences}
\title{Score Sequences with PWM}
\usage{
scoreSequences(sequences, pwm)
}
\arguments{
\item{sequences}{list of sequences}

\item{pwm}{position weight matrix}
}
\value{
list of PWM scores for each sequence
}
\description{
C++ implementation of PWM scoring algorithm
}
