% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\docType{class}
\name{SpectrumScore-class}
\alias{SpectrumScore-class}
\alias{.SpectrumScore}
\alias{spectrumAdjRSquared,SpectrumScore-method}
\alias{spectrumAdjRSquared}
\alias{spectrumDegree,SpectrumScore-method}
\alias{spectrumDegree}
\alias{spectrumResiduals,SpectrumScore-method}
\alias{spectrumResiduals}
\alias{spectrumSlope,SpectrumScore-method}
\alias{spectrumSlope}
\alias{spectrumFStatistic,SpectrumScore-method}
\alias{spectrumFStatistic}
\alias{spectrumFStatisticPValue,SpectrumScore-method}
\alias{spectrumFStatisticPValue}
\alias{spectrumConsistencyScore,SpectrumScore-method}
\alias{spectrumConsistencyScore}
\alias{spectrumConsistencyScorePValue,SpectrumScore-method}
\alias{spectrumConsistencyScorePValue}
\alias{spectrumConsistencyScoreN,SpectrumScore-method}
\alias{spectrumConsistencyScoreN}
\alias{show,SpectrumScore-method}
\alias{plot,SpectrumScore,ANY-method}
\alias{plot,SpectrumScore-method}
\title{An S4 class to represent a scored spectrum}
\usage{
spectrumAdjRSquared(object)

\S4method{spectrumAdjRSquared}{SpectrumScore}(object)

spectrumDegree(object)

\S4method{spectrumDegree}{SpectrumScore}(object)

spectrumResiduals(object)

\S4method{spectrumResiduals}{SpectrumScore}(object)

spectrumSlope(object)

\S4method{spectrumSlope}{SpectrumScore}(object)

spectrumFStatistic(object)

\S4method{spectrumFStatistic}{SpectrumScore}(object)

spectrumFStatisticPValue(object)

\S4method{spectrumFStatisticPValue}{SpectrumScore}(object)

spectrumConsistencyScore(object)

\S4method{spectrumConsistencyScore}{SpectrumScore}(object)

spectrumConsistencyScorePValue(object)

\S4method{spectrumConsistencyScorePValue}{SpectrumScore}(object)

spectrumConsistencyScoreN(object)

\S4method{spectrumConsistencyScoreN}{SpectrumScore}(object)

\S4method{show}{SpectrumScore}(object)

\S4method{plot}{SpectrumScore,ANY}(x)
}
\arguments{
\item{object}{SpectrumScore object}

\item{x}{SpectrumScore object}
}
\value{
Object of type SpectrumScore
}
\description{
An S4 class to represent a scored spectrum

Getter Method spectrumAdjRSquared

Getter Method spectrumDegree

Getter Method spectrumResiduals

Getter Method spectrumSlope

Getter Method spectrumFStatistic

Getter Method spectrumFStatisticPValue

Getter Method spectrumConsistencyScore

Getter Method spectrumConsistencyScorePValue

Getter Method spectrumConsistencyScoreN
}
\section{Slots}{

\describe{
\item{\code{adj_r_squared}}{adjusted \eqn{R^2} of polynomial model}

\item{\code{degree}}{degree of polynomial (integer between 0 and 5)}

\item{\code{residuals}}{residuals of the polynomial model}

\item{\code{slope}}{coefficient of the linear term of the polynomial model
(spectrum "direction")}

\item{\code{f_statistic}}{F statistic from the F test used to determine the
degree of the polynomial model}

\item{\code{f_statistic_p_value}}{p-value associated with the F statistic}

\item{\code{consistency_score}}{raw local consistency score of the spectrum}

\item{\code{consistency_score_p_value}}{p-value associated with the local consistency
score}

\item{\code{consistency_score_n}}{number of permutations performed to calculate
p-value of local consistency score (permutations performed before early
stopping criterion reached)}

\item{\code{plot}}{spectrum plot}
}}

\examples{
new("SpectrumScore",
    adj_r_squared = 0,
    degree = 0L,
    residuals = 0,
    slope = 0,
    f_statistic = 0,
    f_statistic_p_value = 1,
    consistency_score = 1,
    consistency_score_p_value = 1,
    consistency_score_n = 1000L,
    plot = NULL
)
}
