% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\docType{class}
\name{SpectrumScore-class}
\alias{SpectrumScore-class}
\alias{.SpectrumScore}
\alias{adj.r.squared,SpectrumScore-method}
\alias{adj.r.squared}
\alias{degree,SpectrumScore-method}
\alias{degree}
\alias{slope,SpectrumScore-method}
\alias{slope}
\alias{f.statistic,SpectrumScore-method}
\alias{f.statistic}
\alias{f.statistic.p.value,SpectrumScore-method}
\alias{f.statistic.p.value}
\alias{consistency.score,SpectrumScore-method}
\alias{consistency.score}
\alias{consistency.score.p.value,SpectrumScore-method}
\alias{consistency.score.p.value}
\alias{consistency.score.n,SpectrumScore-method}
\alias{consistency.score.n}
\alias{show,SpectrumScore-method}
\alias{residuals,SpectrumScore-method}
\alias{plot,SpectrumScore,ANY-method}
\alias{plot,SpectrumScore-method}
\title{An S4 class to represent a scored spectrum}
\usage{
adj.r.squared(object)

\S4method{adj.r.squared}{SpectrumScore}(object)

degree(object)

\S4method{degree}{SpectrumScore}(object)

slope(object)

\S4method{slope}{SpectrumScore}(object)

f.statistic(object)

\S4method{f.statistic}{SpectrumScore}(object)

f.statistic.p.value(object)

\S4method{f.statistic.p.value}{SpectrumScore}(object)

consistency.score(object)

\S4method{consistency.score}{SpectrumScore}(object)

consistency.score.p.value(object)

\S4method{consistency.score.p.value}{SpectrumScore}(object)

consistency.score.n(object)

\S4method{consistency.score.n}{SpectrumScore}(object)

\S4method{show}{SpectrumScore}(object)

\S4method{residuals}{SpectrumScore}(object)

\S4method{plot}{SpectrumScore,ANY}(x)
}
\arguments{
\item{object}{SpectrumScore object}

\item{x}{SpectrumScore object}
}
\value{
Object of type SpectrumScore
}
\description{
An S4 class to represent a scored spectrum

Getter Method adj.r.squared

Getter Method degree

Getter Method slope

Getter Method f.statistic

Getter Method f.statistic.p.value

Getter Method consistency.score

Getter Method consistency.score.p.value

Getter Method consistency.score.n
}
\section{Slots}{

\describe{
\item{\code{adj.r.squared}}{adjusted \eqn{R^2} of polynomial model}

\item{\code{degree}}{degree of polynomial (integer between 0 and 5)}

\item{\code{residuals}}{residuals of the polynomial model}

\item{\code{slope}}{coefficient of the linear term of the polynomial model
(spectrum "direction")}

\item{\code{f.statistic}}{F statistic from the F test used to determine the
degree of the polynomial model}

\item{\code{f.statistic.p.value}}{p-value associated with the F statistic}

\item{\code{consistency.score}}{raw local consistency score of the spectrum}

\item{\code{consistency.score.p.value}}{p-value associated with the local consistency
score}

\item{\code{consistency.score.n}}{number of permutations performed to calculate
p-value of local consistency score (permutations performed before early
stopping criterion reached)}

\item{\code{plot}}{spectrum plot}
}}

\examples{
new("SpectrumScore", adj.r.squared = 0,
    degree = 0L,
    residuals = 0,
    slope = 0,
    f.statistic = 0,
    f.statistic.p.value = 1,
    consistency.score = 1,
    consistency.score.p.value = 1,
    consistency.score.n = 1000L,
    plot = NULL
)
}
